내 서재에 추가는 완료

+++ src/main/java/toolguys/library/library/dto/user/BookMyFavoriteDTO.java
@@ -7,5 +7,5 @@ public class BookMyFavoriteDTO {
 	private String bookTitle;
 	private String bookWriter;
 	private String bookPub;
-	private int info;
+	private long info;
 }
